requirements/production.txt
amqp==2.6.1
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
Django==3.0.8
django-allauth==0.42.0
django-debug-toolbar==2.2
django-extensions==3.0.5
django-object-actions==2.0.0
django-silk==4.1.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
factory-boy==2.12.0
Faker==4.1.1
gprof2dot==2019.11.30
idna==2.10
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==1.1.1
oauthlib==3.1.0
paramiko==2.7.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
texttable==1.6.2
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.10
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
